refactor: move campaign stats to ops home
